backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-openai==0.1.9
langchain-text-splitters==0.2.0
faiss-cpu
pypdf
docx2txt
tiktoken
pymupdf